Implement LSTM model
Implemented an LSTM model, based on the Keras tutorial:

https://blog.keras.io/a-ten-minute-introduction-to-sequence-to-sequence
-learning-in-keras.html

The LSTM cell takes in individual characters from the story, which is
probably inferior to taking words. Furthermore, this strategy only
predicts the fifth sentence right now, but it doesn't compare the
predicted sentence to the two sentences offered. A method should be
found for picking one of the options.